Simplify, combine components and remove unused
