+++ src/components/Common/StatusCard.tsx
@@ -11,22 +11,18 @@ import classnames, {
   width,
 } from 'classnames/tailwind'
 
-const block = (transparent?: boolean) =>
-  classnames(
-    display('flex'),
-    backgroundColor({ 'bg-primary-background': !transparent }),
-    flexDirection('flex-col'),
-    padding('p-6'),
-    alignItems('items-center'),
-    justifyContent('justify-between'),
-    borderRadius('rounded-lg'),
-    gap('gap-y-4'),
-    width('w-full')
-  )
+const block = classnames(
+  display('flex'),
+  backgroundColor('bg-primary-background'),
+  flexDirection('flex-col'),
+  padding('p-6'),
+  alignItems('items-center'),
+  justifyContent('justify-between'),
+  borderRadius('rounded-lg'),
+  gap('gap-y-4'),
+  width('w-full')
+)
 
-export default function ({
-  transparent,
-  children,
-}: ChildrenProp & { transparent?: boolean }) {
-  return <div className={block(transparent)}>{children}</div>
+export default function ({ children }: ChildrenProp) {
+  return <div className={block}>{children}</div>
 }
